requirements.txt
python-pptx==0.6.23
requests==2.31.0
selenium==4.15.1
python-dotenv==0.19.0
pytest-playwright==1.51.0
# pyderman==2.0.2
# auto-selenium==1.0.1
